requirements.txt
Django
oauth
oauth2
oauth2client
oauthlib
python-social-auth[django]
requests
requests-oauth2
requests-oauthlib
